working on vehicle.js issue
